airflow/requirements.txt
requests
beautifulsoup4
ipykernel
selenium
pandas
python-dotenv
boto3
snowflake-connector-python